==> capstone_bitcoin_tweets_prediction/code/2-tweets-en-no-dup.ipynb <==
import nltk
import re
import string
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import ssl
import pandas as pd 
import numpy
import matplotlib.pyplot as plt
import seaborn as sns
from langdetect import detect
from textblob import TextBlob

ssl._create_default_https_context = ssl._create_unverified_context
sid = SentimentIntensityAnalyzer()
nltk.download('vader_lexicon')

from sklearn.feature_extraction import text

tweets = pd.read_csv('../data/tweets.csv',delimiter=';',lineterminator='\n') 
#Context Scrapped from twitters from 2016-01-01 to 2019-03-29, Collecting Tweets containing Bitcoin or BTC Tools used: - Twint - Tweepy
#https://www.kaggle.com/datasets/alaix14/bitcoin-tweets-20160101-to-20190329 
tweets.head() #going to do sentiment analysis on the text data

After years of price fluctuations ranging between $100 and $900, bitcoin finally broke $1,000 again in January 2017. This kicked off a euphoric bull run phase. Prices doubled to $2,000 in mid-May and then skyrocketed to over $19,000 by December.1

#changing datetimeformat
tweets['timestamp'] = pd.to_datetime(tweets['timestamp'])
tweets = tweets.sort_values(by='timestamp', ascending=True)

#taking those from 2016-2018
tw_df = tweets[(tweets['timestamp'] >= '2016-01-01') & (tweets['timestamp'] <= '2017-12-31')]
tw_df.shape

#cleaning up the column names
tw_df.rename(columns={'text\r':'text'},inplace=True)
#drop duplicates based on user and text
tw_df = tw_df.drop_duplicates(subset=['user', 'text']).reset_index()
#taking only timestamp and text
tw_df = tw_df[['timestamp','text']]

tw_df.head()

tw_df.shape

tw_df.info() #there are almost 16M tweets regarding 'btc' or 'bitcoin' from 1st January 2016 to 29th March 2019 but will take from 2016-2018 instead

#removal of non english word
def english_text(df):
    try: 
        return detect(df)
    except:
        return 'Other'

english_text(tw_df['text'][10])

tw_df['lang']=tw_df['text'].apply(lambda x: english_text(x))
tw_df.head()

tw_df = tw_df[tw_df['lang']=='en']
tw_df.shape

tw_df.to_csv('../data/2-tweets_en_nodup.csv')
