==> capstone_bitcoin_tweets_prediction/code/4-tweets-features.ipynb <==
import numpy as np
import pandas as pd
from textblob import TextBlob
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

tw_df = pd.read_csv('../data/3-preprocessed_tweets.csv')
tw_df.drop(columns='Unnamed: 0',inplace=True)
tw_df.head()

tw_df.isnull().sum()

#getting the subjectivity from TextBlob: Subjectivity is the output that lies within [0,1] and refers to personal opinions and judgments where 0 is very objective and 1 is very subjective
tw_df['subjectivity'] = tw_df['text'].apply(lambda x: TextBlob(x).sentiment.subjectivity)
tw_df.head()

tw_df.info()

tw_df.tail() 
#information of current price tend to get 0 for subjectivity but we also observe some like entry 702024 where it is unable to capture the subjectivity on predicting next pump which likely to be subjective

#getting the Vader sentiment analysis: suitable for english words
sentiment_analyser = SentimentIntensityAnalyzer()
tw_df['compound'] = tw_df['text'].apply(lambda x: sentiment_analyser.polarity_scores(x)["compound"])
tw_df['pos_pol'] = tw_df['text'].apply(lambda x: sentiment_analyser.polarity_scores(x)["pos"])    
tw_df['neg_pol'] = tw_df['text'].apply(lambda x: sentiment_analyser.polarity_scores(x)["neg"])    
tw_df['neu_pol'] = tw_df['text'].apply(lambda x: sentiment_analyser.polarity_scores(x)["neu"])
tw_df.head()

tw_df = tw_df[tw_df['neu_pol']!=1] #those with neutral polarity means that it has no bias opinion, therefore, excluding these data from 

tw_df.head()

tw_df.info()

tw_df['timestamp'] = pd.to_datetime(tw_df['timestamp'])
tw_d = tw_df.resample('D', on='timestamp').agg({'subjectivity':'mean', 'compound':'mean','pos_pol':'mean','neg_pol':'mean','neu_pol':'mean'})
tw_d.to_csv('../data/4-tweets-polarity-day.csv')
tw_d.head()

tw_h = tw_df.resample('H', on='timestamp').agg({'subjectivity':'mean', 'compound':'mean','pos_pol':'mean','neg_pol':'mean','neu_pol':'mean'})
tw_h.to_csv('../data/4-tweets-polarity-hour.csv')
tw_h.head()
