==> capstone_bitcoin_tweets_prediction/code/5-tweets-volume.ipynb <==
import pandas as pd
import numpy as np

tw_df = pd.read_csv('../data/3-preprocessed_tweets.csv')
tw_df.drop(columns='Unnamed: 0',inplace=True)
tw_df['timestamp'] = pd.to_datetime(tw_df['timestamp'])
tw_df.head()

tw_df.info()

#creating volume of tweet features
tw_vol_d = pd.DataFrame(tw_df.resample('D', on='timestamp').text.count())
tw_vol_d.rename(columns={'text':'tweet_volume'})
print(f'There are {tw_vol_d.shape[0]} rows for tweet volume datetime index')
tw_vol_d.head()

tw_vol_d.to_csv('../data/5-tweets-volume-day.csv')

tw_vol_h = pd.DataFrame(tw_df.resample('H', on='timestamp').text.count())
tw_vol_h.rename(columns={'text':'tweet_volume'})
print(f'There are {tw_vol_h.shape[0]} rows for tweet volume datetime index')
tw_vol_h.head()

tw_vol_h.to_csv('../data/5-tweets-volume-hour.csv')
