==> capstone_bitcoin_tweets_prediction/code/6-final-tweet.ipynb <==
import numpy as np
import pandas as pd

tw_pol_d = pd.read_csv('../data/4-tweets-polarity-day.csv')
tw_vol_d = pd.read_csv('../data/5-tweets-volume-day.csv')
tw_pol_h = pd.read_csv('../data/4-tweets-polarity-hour.csv')
tw_vol_h = pd.read_csv('../data/5-tweets-volume-hour.csv')

tw_pol_d.head()

tw_vol_d.head()

tw_pol_h.head()

tw_vol_h.head()

tw_d = pd.merge(tw_pol_d, tw_vol_d, how='left', on='timestamp')
tw_d.to_csv('../data/6-final-tweet-day.csv')
tw_d.head()

tw_d.describe()

tw_h = pd.merge(tw_pol_h, tw_vol_h, how='left', on='timestamp')
tw_h.to_csv('../data/6-final-tweet-hour.csv')
tw_h.head()
