==> capstone_bitcoin_tweets_prediction/code/7-final-dataset.ipynb <==
import pandas as pd
import numpy as np

btc = pd.read_csv('../data/1-bitcoin_data.csv')
tw_d = pd.read_csv('../data/6-final-tweet-day.csv')
tw_h = pd.read_csv('../data/6-final-tweet-hour.csv')
data = [btc,tw_d,tw_h]
for item in data:
    item.drop(columns='Unnamed: 0',inplace=True)
    item['timestamp'] = pd.to_datetime(item['timestamp'])

tw_d['timestamp'] = pd.to_datetime(tw_d['timestamp']).dt.tz_localize(None)
tw_h['timestamp'] = pd.to_datetime(tw_h['timestamp']).dt.tz_localize(None)


#concatinate btc prices and tweets on timestamp
df_d = tw_d.merge(btc, how="left", on="timestamp")
df_h = tw_h.merge(btc, how="left", on="timestamp")
df_d.dropna(axis=0, subset=['Close'],inplace=True)
df_h.dropna(axis=0, subset=['Close'],inplace=True)
df_d.to_csv('../data/7-final-data-day.csv')
df_h.to_csv('../data/7-final-data-hour.csv')
df_h.head()

df_d.head()

df_d.info()

df_h.shape

df_h.info()

df_h.head()

df_h.isnull().sum() #probably during those hours we have only neutral polarity, impute positive pol,neg,compoubd as 0 and neu, neu_pol as 1

df_h['pos_pol'].fillna(0,inplace=True)
df_h['neg_pol'].fillna(0,inplace=True)
df_h['neu_pol'].fillna(1,inplace=True)
df_h['compound'].fillna(0,inplace=True)
df_h['text'].fillna(0,inplace=True)
